requirements.txt
aiohttp==3.7.4.post0
asyncpg==0.22.0
autoflake==1.4
black==20.8b1
databases==0.4.3
dblpy==0.4.0
discoin==4.0
discord.py==1.6.0
jikanpy==4.3.1
owoify-py==1.1.2
Pillow==8.1.0
psutil==5.8.0
psycopg2-binary==2.8.6
pyflakes==2.3.1
pylint==2.6.0
PyNaCl==1.4.0
python-dotenv==0.17.0
uvloop==0.15.2